+++ spell-framework/src/dbus_window_state.rs
@@ -1,11 +1,4 @@
 use crate::{dbus_window_state::second_client::open_internal_clinet, wayland_adapter::SpellWin};
-use smithay_client_toolkit::{
-    reexports::client::protocol::{wl_keyboard, wl_pointer},
-    seat::{
-        keyboard::KeyboardData,
-        pointer::{PointerData, cursor_shape::CursorShapeManager},
-    },
-};
 use std::{
     any::Any,
     future::pending,
@@ -18,19 +11,6 @@ use zbus::{
 };
 
 mod second_client;
-#[derive(Debug)]
-pub struct PointerState {
-    pub pointer: Option<wl_pointer::WlPointer>,
-    pub pointer_data: Option<PointerData>,
-    pub cursor_shape: CursorShapeManager,
-}
-
-#[derive(Debug)]
-pub struct KeyboardState {
-    pub board: Option<wl_keyboard::WlKeyboard>,
-    pub board_data: Option<KeyboardData<SpellWin>>,
-}
-
 /// This a boilerplate trait for connection with CLI, it will be replaced by a procedural
 /// macro in the future.
 pub trait ForeignController: Send + Sync {